+++ js/constants.js
@@ -1,0 +1,8 @@
+const ranges = ()=> ({
+  PHOTO_RANGE: { MIN: 1, MAX: 25 },
+  LIKES_RANGE: { MIN: 15, MAX: 200 },
+  COMMENTS_RANGE: { MIN: 0, MAX: 30 },
+  AVATAR_RANGE: { MIN: 1, MAX: 6 }
+});
+
+export { ranges };
