+++ js/data.js
@@ -1,0 +1,40 @@
+const data = ()=> ({
+  ALTERNATE_TEXTS: [
+    'Вид на пляж с высоты птичьего полета',
+    'Табличка "Как пройти на пляж"',
+    'Море, камни, песок',
+    'Девушка в купальнике с фотоаппаратом',
+    'Две миски с рисом в виде человечков',
+    'Черный спортивный автомобиль с открытой наверх водительской дверью',
+    'Деревянная миска с разрезанной на двое клубникой и вилка',
+    'Две кружки с морсом из красной смородины и ветки смородины',
+    'Девушка на пляже и пролетающий над ней самолет',
+    'Разная обувь в обувнице',
+    'Песчаная дорога на пляж вдоль забора',
+    'Белый спортивный автомобиль "Ауди"',
+    'Блюдо из свежих фруктов и овощей',
+    'Шуточная картинка с рыжим котом',
+    'Фотография ног в тапках в виде ботинок робота',
+    'Летящий самолёт',
+    'Хор, барабанщик и их руководитель на сцене',
+    'Красный ретро автомобиль в помещении из кирпича',
+    'Фотография ног в тапках с фонариком',
+    'Вечерняя фотография пальм на площади с подсветкой',
+    'Мясное кушанье на деревянной тарелке и вилка',
+    'Купающиеся в море пара человек смотрят на закат',
+    'Краб в естественной среде на суше',
+    'Фотография концерта',
+    'Бегемот пытается ухватить автомобиль'
+  ],
+  COMMENT_MESSAGES: [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ],
+  USER_NAMES: ['Артём', 'Крокодил Гена', 'ЯнаЯ', 'Батя', 'Yoshi', 'Yla123']
+});
+
+export { data };
